some apple fixes

+++ src/ggml-backend.c
@@ -205,7 +205,7 @@ static void ggml_backend_cpu_free(ggml_backend_t backend) {
 static const size_t TENSOR_ALIGNMENT = 256; // should be enough for AVX 512 // 256 is the new value - enough for webgpu alignment requirements
 
 static void * ggml_backend_cpu_buffer_get_base(ggml_backend_buffer_t buffer) {
-    size_t base = (void *)(buffer->context);
+    size_t base = (size_t)(void *)(buffer->context);
     size_t addr_incr = 0;
     if (base % TENSOR_ALIGNMENT) {
         addr_incr = TENSOR_ALIGNMENT - (base % TENSOR_ALIGNMENT);
